Remove _impl::make_signed from num_traits.h
- replaced with to _impl::set_signedness in elastic_integer.h
- _impl::common_signedness also moved
